Updating group entity to include join_group features (group code, secure random is imported to allow user to generate random code), and also moved groupdataaccessinterface, to start working on data persistence for groups
